initial_setup + aliases: Setup 'c' to launch calendar
Use 'c' to launch a GUI calendar using zenity. Not using a simple alias
for this, and the expected launch is via dmenu.

I could also choose to source all my aliases into dmenu using a modified
version of dmenu_run_aliases but it is probably not worth the hassle as
there really aren't many aliases that I'd want to run via dmenu.

Also remove the alias for 'c' so that it can instead be used to launch
the script that launches the zenity calendar.