Start a new SSH config file
Move the old SSH config to a backup file to be removed at some point in
the future.

For now, add a single `AddKeysToAgent` property in the new SSH config
file. This property instructs the SSH client to always add the key to a
running agent, so there’s no need to ssh-add it beforehand.